Rename `Prompt` to `Template`
The `Prompt` class really represents a prompt *template* and not a
prompt. I thus renamed it to `Template` to make it less confusing to the
user.
