--- docs/api/prompts.md
@@ -1,1 +1,0 @@
-::: outlines.prompts
