+++ docs/api/templates.md
@@ -1,0 +1,1 @@
+::: outlines.templates
