+++ outlines/__init__.py
@@ -8,7 +8,7 @@ import outlines.types
 from outlines.base import vectorize
 from outlines.caching import clear_cache, disable_cache, get_cache
 from outlines.function import Function
-from outlines.prompts import Prompt, prompt
+from outlines.templates import Template, prompt
 
 __all__ = [
     "clear_cache",
